+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Breaking:
   - Chained queries like `collection.find().where('x').eq('foo')` moved out of the core module into the query-builder plugin
   - The internal `hash()` function does now use a RxDB specific salt
   - Renamed `RxSchema.jsonID` to `RxSchema.jsonSchema`
+  - Moved remaining stuff of leader-election from core into the plugin
 
 Other:
   - Removed many runtime type checks that now should be covered by typescript in buildtime